fix(validator): resolve SQL type mismatch for GPU memory queries
* Fix SQL type mismatch by reading gpu_memory_gb directly as f64 instead of i64
* Update COALESCE default values from 0 to 0.0 for proper REAL type handling
* Remove unnecessary type casting that caused database query failures
* Correct warning message to show 0.0 instead of 0 for consistency

+++ crates/basilica-validator/src/persistence/simple_persistence.rs
@@ -1800,8 +1800,8 @@ impl SimplePersistence {
         executor_id: &str,
         index: u32,
     ) -> Result<f64, anyhow::Error> {
-        let memory: i64 = sqlx::query_scalar(
-            "SELECT COALESCE(gpu_memory_gb, 0) FROM gpu_uuid_assignments
+        let memory: f64 = sqlx::query_scalar(
+            "SELECT COALESCE(gpu_memory_gb, 0.0) FROM gpu_uuid_assignments
              WHERE miner_id = ? AND executor_id = ? AND gpu_index = ?",
         )
         .bind(miner_id)
@@ -1810,7 +1810,7 @@ impl SimplePersistence {
         .fetch_one(&self.pool)
         .await?;
 
-        Ok(memory as f64)
+        Ok(memory)
     }
 
     /// Get the actual gpu_memory_gb for a specific GPU index of an executor from gpu_uuid_assignments
@@ -1820,8 +1820,8 @@ impl SimplePersistence {
         executor_id: &str,
         gpu_uuid: &str,
     ) -> Result<f64, anyhow::Error> {
-        let memory: i64 = sqlx::query_scalar(
-            "SELECT COALESCE(gpu_memory_gb, 0) FROM gpu_uuid_assignments
+        let memory: f64 = sqlx::query_scalar(
+            "SELECT COALESCE(gpu_memory_gb, 0.0) FROM gpu_uuid_assignments
              WHERE miner_id = ? AND executor_id = ? AND gpu_uuid = ?",
         )
         .bind(miner_id)
@@ -1830,7 +1830,7 @@ impl SimplePersistence {
         .fetch_one(&self.pool)
         .await?;
 
-        Ok(memory as f64)
+        Ok(memory)
     }
 
     /// Get the actual gpu_memory_gb for the first GPU (index 0) of an executor from gpu_uuid_assignments
@@ -1839,8 +1839,8 @@ impl SimplePersistence {
         miner_id: &str,
         executor_id: &str,
     ) -> Result<f64, anyhow::Error> {
-        let memory: i64 = sqlx::query_scalar(
-            "SELECT COALESCE(gpu_memory_gb, 0) FROM gpu_uuid_assignments
+        let memory: f64 = sqlx::query_scalar(
+            "SELECT COALESCE(gpu_memory_gb, 0.0) FROM gpu_uuid_assignments
              WHERE miner_id = ? AND executor_id = ? AND gpu_index = 0",
         )
         .bind(miner_id)
